pyproject.toml
[tool.poetry]
name = "inventory"
version = "0.1.0"
description = "Inventory app for Project"
authors = ["Cybertronian123 <atrayanmukherjee0248@gmail.com>"]
license = "MIT"
readme = "README.rst"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-filter = "^24.3"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"
redis = "^5.1.1"
djangorestframework-simplejwt = "^5.3.1"
django-redis = "^5.4.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
